New feature: wait/notify support

+++ extras/test/t01_sched_switch.c
@@ -1,6 +1,6 @@
 /*
  * Copyright (c) 2020 Piotr Stolarz
- * Lightweight Cooperative Threads library
+ * Lightweight cooperative threads library
  *
  * Distributed under the 2-clause BSD License (the License)
  * see accompanying file LICENSE for details.
@@ -28,21 +28,21 @@ static void thrd_proc(void *arg)
 
 int main(int argc, char *argv[])
 {
-    coop_sched_thread(thrd_proc, "thrd_1", 0, (void*)1);
-    coop_sched_thread(thrd_proc, "thrd_2", 0, (void*)2);
-    coop_sched_thread(thrd_proc, "thrd_3", 0, (void*)3);
-    coop_sched_thread(thrd_proc, "thrd_4", 0, (void*)4);
-    coop_sched_thread(thrd_proc, "thrd_5", 0, (void*)5);
+    coop_sched_thread(thrd_proc, "thrd_1", 0, (void*)(size_t)1);
+    coop_sched_thread(thrd_proc, "thrd_2", 0, (void*)(size_t)2);
+    coop_sched_thread(thrd_proc, "thrd_3", 0, (void*)(size_t)3);
+    coop_sched_thread(thrd_proc, "thrd_4", 0, (void*)(size_t)4);
+    coop_sched_thread(thrd_proc, "thrd_5", 0, (void*)(size_t)5);
     coop_sched_service();
 
-    coop_sched_thread(thrd_proc, "thrd_6", 0, (void*)6);
-    coop_sched_thread(thrd_proc, "thrd_7", 0, (void*)5);
-    coop_sched_thread(thrd_proc, "thrd_8", 0, (void*)4);
-    coop_sched_thread(thrd_proc, "thrd_9", 0, (void*)3);
-    coop_sched_thread(thrd_proc, "thrd_10", 0, (void*)2);
+    coop_sched_thread(thrd_proc, "thrd_6", 0, (void*)(size_t)6);
+    coop_sched_thread(thrd_proc, "thrd_7", 0, (void*)(size_t)5);
+    coop_sched_thread(thrd_proc, "thrd_8", 0, (void*)(size_t)4);
+    coop_sched_thread(thrd_proc, "thrd_9", 0, (void*)(size_t)3);
+    coop_sched_thread(thrd_proc, "thrd_10", 0, (void*)(size_t)2);
     coop_sched_service();
 
-    coop_sched_thread(thrd_proc, "thrd_11", 0, (void*)1);
+    coop_sched_thread(thrd_proc, "thrd_11", 0, (void*)(size_t)1);
     coop_sched_service();
 
     return 0;
